Extract shared constrained product builder with Polars support
