+++ baybe/searchspace/utils.py
@@ -2,11 +2,12 @@
 
 from __future__ import annotations
 
-from collections.abc import Sequence
+from collections.abc import Collection, Sequence
 from typing import TYPE_CHECKING
 
 import pandas as pd
 
+from baybe.constraints import DISCRETE_CONSTRAINTS_FILTERING_ORDER
 from baybe.constraints.base import DiscreteConstraint
 from baybe.parameters.base import DiscreteParameter
 
@@ -80,30 +81,33 @@ def optimize_parameter_order(
 
 def parameter_cartesian_prod_polars(
     parameters: Sequence[DiscreteParameter],
+    initial_ldf: pl.LazyFrame | None = None,
 ) -> pl.LazyFrame:
     """Create the Cartesian product of discrete parameter values using Polars.
 
     Args:
         parameters: List of discrete parameter objects.
+        initial_ldf: An optional starting lazy dataframe. When provided, the
+            given parameters are cross-joined into it.
 
     Returns:
         A lazy dataframe containing all possible discrete parameter value combinations.
     """
     from baybe._optional.polars import polars as pl
 
-    if not parameters:
-        return pl.LazyFrame()
-
     # Convert each parameter to a lazy dataframe for cross-join operation
     param_frames = [pl.LazyFrame({p.name: p.active_values}) for p in parameters]
 
-    # Handling edge cases
-    if len(param_frames) == 1:
-        return param_frames[0]
+    # Determine the starting frame
+    if initial_ldf is not None:
+        res = initial_ldf
+    elif param_frames:
+        res = param_frames.pop(0)
+    else:
+        return pl.LazyFrame()
 
-    # Cross-join parameters
-    res = param_frames[0]
-    for frame in param_frames[1:]:
+    # Cross-join remaining parameter frames
+    for frame in param_frames:
         res = res.join(frame, how="cross", force_parallel=True)
 
     return res
@@ -214,3 +218,121 @@ def parameter_cartesian_prod_pandas_constrained(
     df.reset_index(drop=True, inplace=True)
 
     return df
+
+
+def _apply_constraint_filter_pandas(
+    df: pd.DataFrame, constraints: Collection[DiscreteConstraint]
+) -> pd.DataFrame:
+    """Remove discrete search space entries based on constraints.
+
+    The filtering is done inplace, but the modified object is still returned.
+
+    Args:
+        df: The data in experimental representation to be modified inplace.
+        constraints: List of discrete constraints.
+
+    Returns:
+        The filtered dataframe.
+    """
+    # Remove entries that violate parameter constraints:
+    for constraint in (c for c in constraints if c.eval_during_creation):
+        idxs = constraint.get_invalid(df)
+        df.drop(index=idxs, inplace=True)
+    df.reset_index(inplace=True, drop=True)
+
+    return df
+
+
+def _apply_constraint_filter_polars(
+    ldf: pl.LazyFrame,
+    constraints: Sequence[DiscreteConstraint],
+) -> tuple[pl.LazyFrame, list[bool]]:
+    """Remove discrete search space entries based on constraints.
+
+    Note:
+        This will silently skip constraints that have no Polars implementation.
+
+    Args:
+        ldf: The data in experimental representation to be filtered.
+        constraints: Collection of discrete constraints.
+
+    Returns:
+        A tuple containing
+            * The Polars lazyframe with undesired rows removed
+            * A Boolean mask indicating which constraints have **not** been applied
+    """
+    mask_missing = []
+
+    for c in constraints:
+        try:
+            to_keep = c.get_invalid_polars().not_()
+            ldf = ldf.filter(to_keep)
+            mask_missing.append(False)
+        except NotImplementedError:
+            mask_missing.append(True)
+
+    return ldf, mask_missing
+
+
+def build_constrained_product(
+    parameters: Sequence[DiscreteParameter],
+    constraints: Sequence[DiscreteConstraint],
+    initial_df: pd.DataFrame | None = None,
+) -> pd.DataFrame:
+    """Build a constrained Cartesian product, using Polars when available.
+
+    Partitions constraints by Polars support and builds the product accordingly.
+    Parameters covered by Polars-capable constraints are cross-joined and
+    filtered in Polars first, then the remaining parameters and constraints are
+    handled via incremental pandas filtering.
+
+    Args:
+        parameters: The discrete parameters to combine.
+        constraints: The discrete constraints to apply during construction.
+        initial_df: An optional starting dataframe whose columns count as
+            already available for constraint evaluation.
+
+    Returns:
+        A dataframe containing all valid parameter combinations.
+    """
+    from baybe.settings import active_settings
+
+    constraints = sorted(
+        constraints,
+        key=lambda x: DISCRETE_CONSTRAINTS_FILTERING_ORDER.index(x.__class__),
+    )
+
+    remaining_params = list(parameters)
+    remaining_constraints = list(constraints)
+
+    if active_settings.use_polars_for_constraints:
+        from baybe._optional.polars import polars as pl
+
+        polars_constraints = [c for c in constraints if c.has_polars_implementation]
+
+        # Determine which parameters are needed by Polars-capable constraints
+        polars_param_names: set[str] = set()
+        for c in polars_constraints:
+            polars_param_names.update(c._required_parameters)
+        polars_params = [p for p in parameters if p.name in polars_param_names]
+
+        if polars_params:
+            initial_ldf = (
+                pl.from_pandas(initial_df).lazy() if initial_df is not None else None
+            )
+            lazy_df = parameter_cartesian_prod_polars(
+                polars_params, initial_ldf=initial_ldf
+            )
+            lazy_df, _ = _apply_constraint_filter_polars(lazy_df, polars_constraints)
+            initial_df = lazy_df.collect().to_pandas()
+
+            remaining_params = [
+                p for p in parameters if p.name not in polars_param_names
+            ]
+            remaining_constraints = [
+                c for c in constraints if not c.has_polars_implementation
+            ]
+
+    return parameter_cartesian_prod_pandas_constrained(
+        remaining_params, remaining_constraints, initial_df=initial_df
+    )
