+++ tests/constraints/test_constraints_polars.py
@@ -4,11 +4,9 @@ import pytest
 from pandas.testing import assert_frame_equal
 
 from baybe._optional.info import POLARS_INSTALLED
-from baybe.searchspace.discrete import (
+from baybe.searchspace.utils import (
     _apply_constraint_filter_pandas,
     _apply_constraint_filter_polars,
-)
-from baybe.searchspace.utils import (
     parameter_cartesian_prod_pandas,
     parameter_cartesian_prod_polars,
 )
